requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.14
biopython==1.83
python-multipart==0.0.9
